TreasureQuest
The primary aim of the project is to design an autonomous character that collects treasures on a map filled with various treasures and obstacles according to specific requirements. In this gamified project, an algorithm is designed to enable the character to collect these treasures in the shortest possible time, and object-oriented programming is used for the project hierarchy.